Added the Game setup
Adding in the Udacity game start kit.  It includes a game engine, a
resource manager, and some setup code in the app file, among other bits
and pieces.  My job is to complete the game by getting the player
character to be able to move, detect collisions (resetting if that
occurs), and not being allowed to move out of bounds.  Also, the
enemies need to be code to go across the board in a somewhat random
pattern.