Tests for player movement started
Rearrange some stuff, altered how I was going to handle storing player
location.  Storing the pixels and leveraging numberOfRows and Columns
and the rowSize to compute.  Also spent WAY too much time trying to fix
a Sinon-Chai problem until I realized it was a compatibility issue with
the version of Chai I had.